🐛 Handle multiple phrases in parentheses correctly
I noticed during a show that the title “The Wave (Is Coming) (York Back To The
Roots Mix)” was mangled by a comma that was seemingly-randomly inserted into
it, and it was also not possible to remove that comma because something kept
reinserting a comma.

String.join() uses a comma as default value when no joiner is given.