🚀 Allow using Docker volumes for frontend packaging
Okay, this probably warrents a bit of an explanation. I want to be able
to mount the configuration for the frontend into the running container;
however, the configuration is already used during the frontend build and
Docker does not allow using volumes during build.

The workaround here is to not build the frontend during the image build,
but do it as part of starting the container. It feels incredibly dirty
to have to do it like that but… I don’t know, Docker feels just
incredibly immature, built by people who have no idea how software
works, or what people would actually like to do with Docker. IMHO.